+++ src/lib/state/eventStore.ts
@@ -33,6 +33,19 @@ class EventStore {
   }
 
   /**
+   * Removes the argued event by ID if all reminders have occurred or
+   * scheduled date is passed
+   * @param {string} id
+   * @memberof EventStore
+   */
+  removeIfOld(id: string) {
+    const e: CalendarEvent | undefined = this._events.get(id)
+    if (e !== undefined && [...e.reminders.values()].every(v => v)) {
+      this._events.delete(id)
+    }
+  }
+
+  /**
    * Sets the ID in the events cache to the argued CalendarEvent
    * @param {string} id
    * @param {CalendarEvent} e